requirements.txt
#Core libraries
langchain
llama_index
haystack


#For backend server (Optional but recommended)
fastapi
uvicorn

#Additional tools
requests
python-dotenv